refactor: merge folders "manager" and "dataset"
Dataset and draft/commit are not on the same level. DatasetManager and
DraftManager are also not on the same level. To avoid the problem of
circular import, merge them together.

PR Closed: https://github.com/Graviti-AI/graviti-python-sdk/pull/96
