--- graviti/dataset/__init__.py
@@ -1,10 +1,0 @@
-#!/usr/bin/env python3
-#
-# Copyright 2022 Graviti. Licensed under MIT License.
-#
-"""Dataset module."""
-
-from graviti.dataset.dataset import Dataset
-from graviti.dataset.struct import Draft
-
-__all__ = ["Dataset", "Draft"]
